runnable state, discover buggy, imports relative, lots of pylance errors cleared, imports done, more modulization,
