+++ klyqa_ctl/__init__.py
@@ -15,5 +15,5 @@ __version__ = "1.0.11"
 __author__ = 'Frederick Stallmeyer <fraizy@gmx.de>'
 __license__ = 'MIT'
 
-from .klyqa_ctl import Klyqa_account, Data_communicator
+from .klyqa_ctl import *
