+++ klyqa_ctl/devices/light.py
@@ -3,9 +3,11 @@
 from __future__ import annotations
 import argparse
 import json
+
+from general.parameters import add_config_args, get_description_parser
 from .device import *
 
-from ..lib.general import *
+from general.general import *
 
 ## Bulbs ##
 
@@ -184,50 +186,15 @@ BULB_SCENES: list[dict[str, Any]] = [
 class KlyqaBulbResponseStatus(KlyqaDeviceResponse):
     """Klyqa_Bulb_Response_Status"""
 
-    # status: str = ""
-    # color: RGBColor = RGBColor(-1, -1, -1)
-    # brightness: int = -1
-    # temperature: int = -1
-    # active_command: int = -1
-    # active_scene: str = ""
-    # open_slots: int
-    # mode: str = ""
-    # fwversion: str = ""
-    # sdkversion: str = ""
-    # ts: datetime.datetime = datetime.datetime.now()
-
     def __str__(self) -> str:
         """__str__"""
         return get_obj_attr_values_as_string(self)
 
     def __init__(
         self,
-        # type: str,
-        # status: str = "off",
-        # brightness: dict[str, int] = {},
-        # mode: str = "",
-        # open_slots: int = 0,
-        # fwversion: str = "",
-        # sdkversion: str = "",
-        # color: dict[str, int] = {},
-        # temperature: int = -1,
-        # active_command: int = 0,
-        # active_scene: str = 0,
         **kwargs
     ) -> None:
         """__init__"""
-        # self.status = status
-        # self.color = (
-        #     RGBColor(color["red"], color["green"], color["blue"]) if color else {}
-        # )
-        # self.brightness = brightness["percentage"] if brightness else 0
-        # self.temperature = temperature
-        # self.active_command = active_command
-        # self.active_scene = active_scene
-        # self.open_slots = open_slots
-        # self.mode = mode
-        # self.fwversion = fwversion
-        # self.sdkversion = sdkversion
         self.active_command: int | None = None
         self.active_scene: str | None = None
         self._brightness: int | None = None
@@ -271,13 +238,14 @@ class KlyqaBulb(KlyqaDevice):
     def setTemp(self, temp: int):
         temperature_enum = []
         try:
-            temperature_enum = [
-                trait["value_schema"]["properties"]["colorTemperature"]["enum"]
-                for trait in device_configs[self.ident.product_id]["deviceTraits"]
-                if trait["trait"] == "@core/traits/color-temperature"
-            ]
-            if len(temperature_enum) < 2:
-                raise Exception()
+            if self.ident:
+                temperature_enum = [
+                    trait["value_schema"]["properties"]["colorTemperature"]["enum"]
+                    for trait in device_configs[self.ident.product_id]["deviceTraits"]
+                    if trait["trait"] == "@core/traits/color-temperature"
+                ]
+                if len(temperature_enum) < 2:
+                    raise Exception()
         except:
             LOGGER.error("No temperature change on the bulb available")
             return False
@@ -532,3 +500,591 @@ def add_command_args_bulb(parser: argparse.ArgumentParser) -> None:
     parser.add_argument("--mystic", help="Mystic Mountain", action="store_true")
     parser.add_argument("--cotton", help="Cotton Candy", action="store_true")
     parser.add_argument("--ice", help="Ice Cream", action="store_true")
+
+
+async def process_args_to_msg_lighting(args, args_in, send_to_devices_cb, message_queue_tx_local, message_queue_tx_command_cloud, scene_list: list[str]) -> bool:
+    """process_args_to_msg_lighting"""
+    
+    def local_and_cloud_command_msg(json_msg, timeout) -> None:
+        message_queue_tx_local.append((json.dumps(json_msg), timeout))
+        message_queue_tx_command_cloud.append(json_msg)
+        
+    # TODO: Missing cloud discovery and interactive device selection. Send to devices if given as argument working.
+    if (args.local or args.tryLocalThanCloud) and (
+        not args.device_name
+        and not args.device_unitids
+        and not args.allDevices
+        and not args.discover
+    ):
+        discover_local_args: list[str] = [
+            DeviceType.lighting.name,
+            "--request",
+            "--allDevices",
+            "--selectDevice",
+            "--discover"
+        ]
+
+        orginal_args_parser: argparse.ArgumentParser = get_description_parser()
+        discover_local_args_parser: argparse.ArgumentParser = get_description_parser()
+
+        add_config_args(parser=orginal_args_parser)
+        add_config_args(parser=discover_local_args_parser)
+        add_command_args_bulb(parser=discover_local_args_parser)
+
+        original_config_args_parsed, _ = orginal_args_parser.parse_known_args(
+            args=args_in
+        )
+
+        discover_local_args_parsed = discover_local_args_parser.parse_args(
+            discover_local_args, namespace=original_config_args_parsed
+        )
+
+        uids = await send_to_devices_cb(
+            discover_local_args_parsed
+        )
+        if isinstance(uids, set) or isinstance(uids, list):
+            # args_in.extend(["--device_unitids", ",".join(list(uids))])
+            args_in = ["--device_unitids", ",".join(list(uids))] + args_in
+        elif isinstance(uids, str) and uids == "no_devices":
+            return False
+        else:
+            LOGGER.error("Error during local discovery of the devices.")
+            return False
+
+        add_command_args_bulb(parser=orginal_args_parser)
+        args = orginal_args_parser.parse_args(args=args_in, namespace=args)
+        
+    commands_to_send: list[str] = [i for i in commands_send_to_bulb if hasattr(args, i) and getattr(args, i)]
+        
+    if commands_to_send:
+        print("Commands to send to devices: " + ", ".join(commands_to_send))
+    else:
+        print("Commands (arguments):")
+        print(sep_width * "-")
+
+        def get_temp_range(product_id):
+            temperature_enum = []
+            try:
+                temperature_enum = [
+                    trait["value_schema"]["properties"]["colorTemperature"][
+                        "enum"
+                    ]
+                    if "properties" in trait["value_schema"]
+                    else trait["value_schema"]["enum"]
+                    for trait in device_configs[product_id]["deviceTraits"]
+                    if trait["trait"] == "@core/traits/color-temperature"
+                ]
+                if len(temperature_enum[0]) < 2:
+                    raise Exception()
+            except:
+                return False
+            return temperature_enum[0]
+
+        def get_inner_range(tup1, tup2) -> tuple[int, int]:
+            return (
+                tup1[0] if tup1[0] > tup2[0] else tup2[0],
+                tup1[1] if tup1[1] < tup2[1] else tup2[1],
+            )
+
+        temperature_enum = []
+        color: list[int] = [0, 255]
+        brightness: list[int] = [0, 100]
+        for u_id in args.device_unitids[0].split(","):
+            u_id: str = format_uid(u_id)
+            if u_id not in self.devices or not self.devices[u_id].ident:
+
+                async def send_ping() -> bool:
+                    discover_local_args2: list[str] = ["--ping", "--device_unitids", u_id]
+
+                    orginal_args_parser: argparse.ArgumentParser = get_description_parser()
+                    discover_local_args_parser2: argparse.ArgumentParser = get_description_parser()
+
+                    add_config_args(parser=orginal_args_parser)
+                    add_config_args(parser=discover_local_args_parser2)
+                    add_command_args_bulb()(parser=discover_local_args_parser2)
+
+                    (
+                        original_config_args_parsed,
+                        _,
+                    ) = orginal_args_parser.parse_known_args(args=args_in)
+
+                    discover_local_args_parsed2 = (
+                        discover_local_args_parser2.parse_args(
+                            discover_local_args2,
+                            namespace=original_config_args_parsed,
+                        )
+                    )
+
+                    ret = await self._send_to_devices(
+                        discover_local_args_parsed2,
+                        args_in,
+                        udp=udp,
+                        tcp=tcp,
+                        timeout_ms=3000,
+                    )
+                    if isinstance(ret, bool) and ret:
+                        return True
+                    else:
+                        return False
+
+                ret: bool = await send_ping()
+                if isinstance(ret, bool) and ret:
+                    product_id: str = self.devices[u_id].ident.product_id
+                else:
+                    LOGGER.error(f"Device {u_id} not found.")
+                    return False
+            product_id = self.devices[u_id].ident.product_id
+            if not temperature_enum:
+                temperature_enum = get_temp_range(product_id)
+            else:
+                temperature_enum: tuple[int, int] = get_inner_range(
+                    temperature_enum, get_temp_range(product_id)
+                )
+        arguments_send_to_device = {}
+        if temperature_enum:
+            arguments_send_to_device = {
+                "temperature": " ["
+                + str(temperature_enum[0])
+                + "-"
+                + str(temperature_enum[1])
+                + "] (Kelvin, low: warm, high: cold)"
+            }
+
+        arguments_send_to_device: dict[str, str] = {
+            **arguments_send_to_device,
+            **{
+                "color": f" rgb [{color[0]}..{color[1]}] [{color[0]}..{color[1]}] [{color[0]}..{color[1]}]",
+                "brightness": " ["
+                + str(brightness[0])
+                + ".."
+                + str(brightness[1])
+                + "]",
+                "routine_commands": " [cmd]",
+                "routine_id": " [int]",
+                "power": " [on/off]",
+            },
+        }
+
+        count = 1
+        for c in commands_send_to_bulb:
+            args_to_b: str = (
+                arguments_send_to_device[c] if c in arguments_send_to_device else ""
+            )
+            print(str(count) + ") " + c + args_to_b)
+            count: int = count + 1
+
+        cmd_c_id: int = int(input("Choose command number [1-9]*: "))
+        if cmd_c_id > 0 and cmd_c_id < count:
+            args_in.append("--" + commands_send_to_bulb[cmd_c_id - 1])
+        else:
+            LOGGER.error("No such command id " + str(cmd_c_id) + " available.")
+            sys.exit(1)
+
+        if commands_send_to_bulb[cmd_c_id - 1] in arguments_send_to_device:
+            args_app: str = input(
+                "Set arguments (multiple arguments space separated) for command [Enter]: "
+            )
+            if args_app:
+                args_in.extend(args_app.split(" "))
+
+        parser: ArgumentParser = get_description_parser()
+
+        add_config_args(parser=parser)
+        add_command_args(parser=parser)
+
+        args = parser.parse_args(args_in, namespace=args)
+        # args.func(args)
+
+    if args.ota is not None:
+        local_and_cloud_command_msg(
+            ({"type": "fw_update", "url": args.ota}, 10000)
+        )
+
+    if args.ping:
+        local_and_cloud_command_msg({"type": "ping"}, 10000)
+
+    if args.request:
+        local_and_cloud_command_msg({"type": "request"}, 10000)
+
+    if args.enable_tb is not None:
+        a = args.enable_tb[0]
+        if a != "yes" and a != "no":
+            print("ERROR --enable_tb needs to be yes or no")
+            sys.exit(1)
+
+        message_queue_tx_local.append(
+            (json.dumps({"type": "backend", "link_enabled": a}), 1000)
+        )
+
+    if args.passive:
+        pass
+
+    if args.power:
+        message_queue_tx_local.append(
+            (json.dumps({"type": "request", "status": args.power}), 500)
+        )
+        message_queue_tx_state_cloud.append({"status": args.power})
+
+    def forced_continue(reason: str) -> bool:
+        if not args.force:
+            LOGGER.error(reason)
+            return False
+        else:
+            LOGGER.info(reason)
+            LOGGER.info("Enforced send")
+            return True
+
+    """range"""
+    Check_device_parameter = Enum(
+        "Check_device_parameter", "color brightness temperature scene"
+    )
+
+    def missing_config(product_id) -> bool:
+        if not forced_continue(
+            "Missing or faulty config values for device "
+            + " product_id: "
+            + product_id
+        ):
+            return True
+        return False
+
+    #### Needing config profile versions implementation for checking trait ranges ###
+
+    def check_color_range(product_id, values) -> bool:
+        color_enum = []
+        try:
+            # # different device trait schematics. for now go typical range
+            # color_enum = [
+            #     trait["value_schema"]["definitions"]["color_value"]
+            #     for trait in device_configs[product_id]["deviceTraits"]
+            #     if trait["trait"] == "@core/traits/color"
+            # ]
+            # color_range = (
+            #     color_enum[0]["minimum"],
+            #     color_enum[0]["maximum"],
+            # )
+            color_range = (
+                0,
+                255,
+            )
+            for value in values:
+                if int(value) < color_range[0] or int(value) > color_range[1]:
+                    return forced_continue(
+                        f"Color {value} out of range [{color_range[0]}..{color_range[1]}]."
+                    )
+
+        except:
+            return not missing_config(product_id)
+        return True
+
+    def check_brightness_range(product_id, value) -> bool:
+        brightness_enum = []
+        try:
+            # different device trait schematics. for now go typical range
+            # brightness_enum = [
+            #     trait["value_schema"]["properties"]["brightness"]
+            #     for trait in device_configs[product_id]["deviceTraits"]
+            #     if trait["trait"] == "@core/traits/brightness"
+            # ]
+            # brightness_range = (
+            #     brightness_enum[0]["minimum"],
+            #     brightness_enum[0]["maximum"],
+            # )
+            brightness_range = (
+                0,
+                100,
+            )
+
+            if (
+                int(value) < brightness_range[0]
+                or int(value) > brightness_range[1]
+            ):
+                return forced_continue(
+                    f"Brightness {value} out of range [{brightness_range[0]}..{brightness_range[1]}]."
+                )
+
+        except:
+            return not missing_config(product_id)
+        return True
+
+    def check_temp_range(product_id, value) -> bool:
+        temperature_range: list[int] = []
+        try:
+            # # different device trait schematics. for now go typical range
+            # temperature_range = [
+            #     trait["value_schema"]["properties"]["colorTemperature"]["enum"]
+            #     if "properties" in trait["value_schema"]
+            #     else trait["value_schema"]["enum"]
+            #     for trait in device_configs[product_id]["deviceTraits"]
+            #     if trait["trait"] == "@core/traits/color-temperature"
+            # ][0]
+            temperature_range = [2000, 6500]
+            if (
+                int(value) < temperature_range[0]
+                or int(value) > temperature_range[1]
+            ):
+                return forced_continue(
+                    f"Temperature {value} out of range [{temperature_range[0]}..{temperature_range[1]}]."
+                )
+        except Exception as excp:
+            return not missing_config(product_id)
+        return True
+
+    def check_scene_support(product_id, scene) -> bool:
+        color_enum = []
+        try:
+            # color_enum = [
+            #     trait
+            #     for trait in device_configs[product_id]["deviceTraits"]
+            #     if trait["trait"] == "@core/traits/color"
+            # ]
+            # color_support = len(color_enum) > 0
+
+            # if not color_support and not "cwww" in scene:
+            #     return forced_continue(
+            #         f"Scene {scene['label']} not supported by device product {product_id}."
+            #     )
+            return True
+
+        except:
+            return not missing_config(product_id)
+        return True
+
+    check_range = {
+        Check_device_parameter.color: check_color_range,
+        Check_device_parameter.brightness: check_brightness_range,
+        Check_device_parameter.temperature: check_temp_range,
+        Check_device_parameter.scene: check_scene_support,
+    }
+
+    def check_device_parameter(
+        parameter: Check_device_parameter, values, product_id
+    ) -> bool:
+        # if not device_configs and not forced_continue("Missing configs for devices."):
+        #     return False
+
+        # for u_id in target_device_uids:
+        #     # dev = [
+        #     #     device["productId"]
+        #     #     for device in self.acc_settings["devices"]
+        #     #     if format_uid(device["localDeviceId"]) == format_uid(u_id)
+        #     # ]
+        #     # product_id = dev[0]
+        #     product_id = self.devices[u_id].ident.product_id
+
+        if not check_range[parameter](product_id, values):
+            return False
+        return True
+
+    if args.color is not None:
+        r, g, b = args.color
+        # if not check_device_parameter(Check_device_parameter.color, [r, g, b]):
+        #     return False
+
+        tt = args.transitionTime[0]
+        msg = color_message(
+            r, g, b, int(tt), skipWait=args.brightness is not None
+        )
+
+        check_color = functools.partial(
+            check_device_parameter, Check_device_parameter.color, [r, g, b]
+        )
+        msg = msg + (check_color,)
+
+        message_queue_tx_local.append(msg)
+        col = json.loads(msg[0])["color"]
+        message_queue_tx_state_cloud.append({"color": col})
+
+    if args.temperature is not None:
+        temperature = args.temperature[0]
+        # if not check_device_parameter(Check_device_parameter.temperature, temperature):
+        #     return False
+
+        tt = args.transitionTime[0]
+        msg = temperature_message(
+            temperature, int(tt), skipWait=args.brightness is not None
+        )
+
+        check_temperature = functools.partial(
+            check_device_parameter, Check_device_parameter.temperature, temperature
+        )
+        msg = msg + (check_temperature,)
+
+        temperature = json.loads(msg[0])["temperature"]
+        message_queue_tx_local.append(msg)
+        message_queue_tx_state_cloud.append({"temperature": temperature})
+
+    if args.brightness is not None:
+        brightness = args.brightness[0]
+        # if not check_device_parameter(Check_device_parameter.brightness, brightness):
+        #     return False
+
+        tt = args.transitionTime[0]
+        msg: tuple[str, int] = brightness_message(brightness, int(tt))
+
+        check_brightness = functools.partial(
+            check_device_parameter, Check_device_parameter.brightness, brightness
+        )
+        msg = msg + (check_brightness,)
+
+        message_queue_tx_local.append(msg)
+        brightness = json.loads(msg[0])["brightness"]
+        message_queue_tx_state_cloud.append({"brightness": brightness})
+
+    if args.percent_color is not None:
+        if not device_configs and not forced_continue(
+            "Missing configs for devices."
+        ):
+            return False
+        r, g, b, w, c = args.percent_color
+        tt = args.transitionTime[0]
+        msg = percent_color_message(
+            r, g, b, w, c, int(tt), skipWait=args.brightness is not None
+        )
+        message_queue_tx_local.append(msg)
+        p_color = json.loads(msg[0])["p_color"]
+        message_queue_tx_state_cloud.append({"p_color": p_color})
+
+    if args.factory_reset:
+        local_and_cloud_command_msg({"type": "factory_reset"}, 500)
+
+    if args.fade is not None and len(args.fade) == 2:
+        local_and_cloud_command_msg({"type": "request","fade_out": args.fade[1],"fade_in": args.fade[0]}, 500)
+
+    scene = ""
+    if args.WW:
+        scene = "Warm White"
+    if args.daylight:
+        scene = "Daylight"
+    if args.CW:
+        scene = "Cold White"
+    if args.nightlight:
+        scene = "Night Light"
+    if args.relax:
+        scene = "Relax"
+    if args.TVtime:
+        scene = "TV time"
+    if args.comfort:
+        scene = "Comfort"
+    if args.focused:
+        scene = "Focused"
+    if args.fireplace:
+        scene = "Fireplace"
+    if args.club:
+        scene = "Jazz Club"
+    if args.romantic:
+        scene = "Romantic"
+    if args.gentle:
+        scene = "Gentle"
+    if args.summer:
+        scene = "Summer"
+    if args.jungle:
+        scene = "Jungle"
+    if args.ocean:
+        scene = "Ocean"
+    if args.fall:  # Autumn
+        scene = "Fall"
+    if args.sunset:
+        scene = "Sunset"
+    if args.party:
+        scene = "Party"
+    if args.spring:
+        scene = "Spring"
+    if args.forest:
+        scene = "Forest"
+    if args.deep_sea:
+        scene = "Deep Sea"
+    if args.tropical:
+        scene = "Tropical"
+    if args.magic:
+        scene = "Magic Mood"
+    if args.mystic:
+        scene = "Mystic Mountain"
+    if args.cotton:
+        scene = "Cotton Candy"
+    if args.ice:
+        scene = "Ice Cream"
+
+    # if (
+    #     args.WW
+    #     or args.daylight
+    #     or args.CW
+    #     or args.nightlight
+    #     or args.relax
+    #     or args.TVtime
+    #     or args.comfort
+    #     or args.focused
+    #     or args.fireplace
+    #     or args.club
+    #     or args.romantic
+    #     or args.gentle
+    #     or args.summer
+    #     or args.jungle
+    #     or args.ocean
+    #     or args.fall  # Autumn
+    #     or args.sunset
+    #     or args.party
+    #     or args.spring
+    #     or args.forest
+    #     or args.deep_sea
+    #     or args.tropical
+    #     or args.magic
+    #     or args.mystic
+    #     or args.cotton
+    #     or args.ice
+    #     # or args.monjito
+    # ):
+    if scene:
+        scene_result = [x for x in BULB_SCENES if x["label"] == scene]
+        if not len(scene_result) or len(scene_result) > 1:
+            LOGGER.error(
+                f"Scene {scene} not found or more than one scene with the name found."
+            )
+            return False
+        scene_obj = scene_result[0]
+
+        # check_brightness = functools.partial(check_device_parameter, Check_device_parameter.scene, scene_obj)
+        # msg = msg + (check_brightness,)
+        # if not check_device_parameter(Check_device_parameter.scene, scene_obj):
+        #     return False
+        commands = scene_obj["commands"]
+        if len(commands.split(";")) > 2:
+            commands += "l 0;"
+        args.routine_id = 0
+        args.routine_put = True
+        args.routine_commands = commands
+        args.routine_scene = str(scene_obj["id"])
+
+    if args.routine_list:
+        local_and_cloud_command_msg({"type": "routine", "action": "list"}, 500)
+
+    if args.routine_put and args.routine_id is not None:
+        local_and_cloud_command_msg(
+            {
+                "type": "routine",
+                "action": "put",
+                "id": args.routine_id,
+                "scene": args.routine_scene,
+                "commands": args.routine_commands,
+            },
+            500,
+        )
+
+    if args.routine_delete and args.routine_id is not None:
+        local_and_cloud_command_msg(
+            {"type": "routine", "action": "delete", "id": args.routine_id}, 500
+        )
+    if args.routine_start and args.routine_id is not None:
+        local_and_cloud_command_msg(
+            {"type": "routine", "action": "start", "id": args.routine_id}, 500
+        )
+
+    if args.reboot:
+        local_and_cloud_command_msg({"type": "reboot"}, 500)
+        
+    if scene:
+        scene_list.append(scene)
+
+
+# async def discover_lightings(args, args_in, send_to_devices_cb):
+
